+++ brutifus/__init__.py
@@ -1,2 +1,2 @@
-from .brutifus import * # So that users only need to do import brutus
-from .brutifus_metadata import __version__
+#from .brutifus import * # So that users only need to do import brutifus
+from .brutifus_version import __version__ # Gives users easy access to the version