+++ brutifus/brutifus_metadata.py
@@ -18,8 +18,6 @@ import matplotlib.pyplot as plt
 
 # ---| Some basic parameters |------------------------------------------------------------
 
-__version__ = '2019.02.3'
-
 # Where are we located ?
 bifus_dir = os.path.dirname(__file__) # Get the project directory from the file location!