+++ brutifus/brutifus_version.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+'''
+brutifus: a set of Python modules to process datacubes from integral field spectrographs.\n
+Copyright (C) 2018-2019,  F.P.A. Vogt
+
+-----------------------------------------------------------------------------------------
+
+This file contains the version of the code.
+
+Created August 2019, F.P.A. Vogt - frederic.vogt@alumni.anu.edu.au
+'''
+# ----------------------------------------------------------------------------------------
+
+__version__ = '2019.08.1'